refactor: create separate table components, rename components
